# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Consize API
  version: 0.3.0-alpha
  description: The contract for the current Consize product and operational HTTP API.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://127.0.0.1:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/dashboard:
    get:
      tags: [Product]
      operationId: getDashboard
      summary: Get the operator dashboard
      description: Returns the current resources, recommendations, plugins, actions, jobs, policy, and summary statistics.
      responses:
        '200': {$ref: '#/components/responses/DashboardResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '500': {$ref: '#/components/responses/ErrorResponse'}
  /api/resources:
    get:
      tags: [Product]
      operationId: listResources
      summary: List registered resources
      description: Lists provider-neutral resources currently held by the registry.
      responses:
        '200': {$ref: '#/components/responses/ResourcesResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '500': {$ref: '#/components/responses/ErrorResponse'}
    post:
      tags: [Product]
      operationId: upsertResource
      summary: Register a resource
      description: Creates or updates one provider-neutral resource; administrator authorization is required.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResourceInput'}
      responses:
        '200':
          description: Resource stored successfully.
          headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
          content: {application/json: {schema: {$ref: '#/components/schemas/Resource'}}}
        '400': {$ref: '#/components/responses/ErrorResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
  /api/discovery:
    post:
      tags: [Product]
      operationId: runDiscovery
      summary: Run configured discovery plugins
      description: Collects provider observations and upserts their normalized resources into the registry.
      responses:
        '200':
          description: Discovery completed.
          headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema:
                type: object
                required: [providers]
                properties:
                  providers: {type: array, items: {$ref: '#/components/schemas/DiscoveryResult'}}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '502': {$ref: '#/components/responses/ErrorResponse'}
  /api/plugins:
    get:
      tags: [Product]
      operationId: listPlugins
      summary: List plugin status
      description: Lists registered plugin manifests with their latest health status.
      responses:
        '200': {$ref: '#/components/responses/PluginsResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '500': {$ref: '#/components/responses/ErrorResponse'}
  /api/actions:
    get:
      tags: [Product]
      operationId: listActions
      summary: List action events
      description: Lists durable action and audit events visible to dashboard users.
      responses:
        '200': {$ref: '#/components/responses/ActionsResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '500': {$ref: '#/components/responses/ErrorResponse'}
  /api/action-records:
    get:
      tags: [Product]
      operationId: listActionRecords
      summary: List durable actions
      description: Lists governed action attempts separately from append-only audit events.
      responses:
        '200': {$ref: '#/components/responses/ActionRecordsResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '500': {$ref: '#/components/responses/ErrorResponse'}
  /api/jobs:
    get:
      tags: [Product]
      operationId: listJobs
      summary: List durable jobs
      description: Lists action jobs, including recovery and verification state.
      responses:
        '200': {$ref: '#/components/responses/JobsResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '503': {$ref: '#/components/responses/ErrorResponse'}
  /api/recommendations/generate:
    post:
      tags: [Product]
      operationId: generateRecommendations
      summary: Generate recommendations
      description: Reads configured metrics for one resource and persists recommendations produced by the active algorithm.
      parameters:
        - name: resource_id
          in: query
          required: true
          description: Stable registry identifier of the resource to evaluate.
          schema: {type: string, minLength: 1}
      responses:
        '200':
          description: Recommendations generated.
          headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema:
                type: object
                required: [recommendations, evidence]
                properties:
                  recommendations: {type: array, items: {$ref: '#/components/schemas/Recommendation'}}
                  evidence: {$ref: '#/components/schemas/MetricsSnapshot'}
        '400': {$ref: '#/components/responses/ErrorResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '404': {$ref: '#/components/responses/ErrorResponse'}
        '502': {$ref: '#/components/responses/ErrorResponse'}
...
  /api/recommendations/{recommendation_id}/execute:
    post:
      tags: [Product]
      operationId: executeRecommendation
      summary: Execute an approved recommendation
      description: Persists an approved durable job for worker execution, verification, and recovery.
      parameters: [{$ref: '#/components/parameters/RecommendationId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {$ref: '#/components/requestBodies/ActionRequest'}
      responses:
        '202': {$ref: '#/components/responses/JobAcceptedResponse'}
        '400': {$ref: '#/components/responses/ErrorResponse'}
        '401': {$ref: '#/components/responses/ErrorResponse'}
        '403': {$ref: '#/components/responses/ErrorResponse'}
        '503': {$ref: '#/components/responses/ErrorResponse'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  headers:
    RequestId:
      description: Correlation identifier generated or echoed by the API.
      schema: {type: string, minLength: 1}
...
  responses:
    ErrorResponse:
      description: The request could not be completed.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}
...
    DashboardResponse:
      description: Current dashboard state.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {$ref: '#/components/schemas/Dashboard'}}}
    ResourcesResponse:
      description: Registered resources.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {type: object, required: [resources], properties: {resources: {type: array, items: {$ref: '#/components/schemas/Resource'}}}}}}
    PluginsResponse:
      description: Registered plugin status.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {type: object, required: [plugins], properties: {plugins: {type: array, items: {$ref: '#/components/schemas/PluginStatus'}}}}}}
    ActionsResponse:
      description: Action events.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {type: object, required: [actions], properties: {actions: {type: array, items: {$ref: '#/components/schemas/ActionEvent'}}}}}}
    ActionRecordsResponse:
      description: Durable governed action attempts.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {type: object, required: [actions], properties: {actions: {type: array, items: {$ref: '#/components/schemas/Action'}}}}}}
    JobsResponse:
      description: Durable action jobs.
      headers: {X-Request-ID: {$ref: '#/components/headers/RequestId'}}
      content: {application/json: {schema: {type: object, required: [jobs], properties: {jobs: {type: array, items: {$ref: '#/components/schemas/Job'}}}}}}
...
  schemas:
...
    Error: {type: object, required: [error], properties: {error: {type: string}}}
...
    Resource:
      type: object
      required: [schema_version, id, type, provider, provider_resource_id, name, environment, owner, region, account, criticality, lifecycle_state, support_status, labels, metadata, current_state, monthly_cost, source_plugin_id, observed_at, first_seen_at, last_seen_at, created_at, updated_at]
      properties:
        schema_version: {type: integer, minimum: 1, description: Version of the universal resource contract.}
        id: {type: string, minLength: 1, description: Stable Consize identity derived only from immutable provider coordinates for new resources.}
        type: {type: string, minLength: 1, description: Provider-neutral resource type.}
        provider: {type: string, minLength: 1}
        provider_resource_id: {type: string, minLength: 1, description: Stable identity assigned by the provider.}
        name: {type: string, minLength: 1}
        environment: {type: string}
        owner: {type: string, minLength: 1}
        region: {type: string, minLength: 1, description: Provider location or explicit non-regional scope.}
        account: {type: string, minLength: 1, description: "Provider account, project, subscription, or cluster identity."}
        criticality: {type: string, enum: [low, medium, high]}
        lifecycle_state: {type: string, enum: [active, stale, deleted]}
        support_status: {type: string, enum: [fully_supported, model_only, unsupported]}
        labels: {type: object, additionalProperties: {type: string}}
        metadata: {$ref: '#/components/schemas/JsonObject'}
        current_state: {$ref: '#/components/schemas/JsonObject'}
        monthly_cost: {type: number}
        source_plugin_id: {type: string}
        observed_at: {type: string, format: date-time}
        first_seen_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ResourceInput:
      type: object
      description: Resource observation accepted from an administrator or trusted discovery adapter. Server-managed model and lifecycle fields are returned in Resource.
      required: [id, type, provider, provider_resource_id, name, environment, owner, region, account, criticality, labels, metadata, current_state, monthly_cost, source_plugin_id, observed_at, created_at, updated_at]
      properties:
        id: {type: string}
        type: {type: string}
        provider: {type: string}
        provider_resource_id: {type: string}
        name: {type: string}
        environment: {type: string}
        owner: {type: string}
        region: {type: string}
        account: {type: string}
        criticality: {type: string, enum: [low, medium, high]}
        labels: {type: object, additionalProperties: {type: string}}
        metadata: {$ref: '#/components/schemas/JsonObject'}
        current_state: {$ref: '#/components/schemas/JsonObject'}
        monthly_cost: {type: number}
        source_plugin_id: {type: string}
        observed_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Recommendation:
      type: object
      required: [id, resource_id, plugin_id, algorithm_id, action_type, title, summary, current, proposed, parameters, estimated_savings_monthly, confidence, risk, evidence, policy_id, status, created_at, updated_at]
      properties:
        schema_version: {type: integer, minimum: 1}
        id: {type: integer, format: int64}
        resource_id: {type: string}
        plugin_id: {type: string}
        algorithm_id: {type: string}
        algorithm_version: {type: string}
        recommendation_type: {type: string}
        action_type: {type: string}
        title: {type: string}
        summary: {type: string}
        current: {$ref: '#/components/schemas/JsonObject'}
        proposed: {$ref: '#/components/schemas/JsonObject'}
        parameters: {$ref: '#/components/schemas/JsonObject'}
        estimated_savings_monthly: {type: number}
        savings_estimate: {$ref: '#/components/schemas/SavingsEstimate'}
        cost_estimate: {$ref: '#/components/schemas/CostEstimate'}
        confidence: {type: string}
        risk: {type: string}
        evidence: {type: array, items: {type: string}}
        evidence_refs: {type: array, minItems: 1, items: {type: string}}
        policy_id: {type: string}
        status: {type: string, description: "Recommendation lifecycle state: pending, planned, approved, executing, verified, rejected, expired, superseded, failed, rolled_back, or manual_intervention."}
        expires_at: {type: string, format: date-time}
        superseded_by: {type: integer, format: int64}
        status_reason: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
...
    DiscoveryResult:
      type: object
      required: [plugin_id, count, resources]
      properties:
        plugin_id: {type: string}
        count: {type: integer}
        resources: {type: array, items: {$ref: '#/components/schemas/Resource'}}